Write empty line at meta end and fix resetting
We write an empty line '\n' at the end of the header, this line can
be omitted in the text version but we have it in for consistency.
In the binary version it ensures that header data won't be interpreted
as a line. We also fix resetting. Prior to this change we would have
resetted before the empty line.